+++ server/rules/clothing.js
